Allow integers and irrationals in Rotation angle
An innocent user trying
    act!(object1, Action(anim_rotate_around(π, object2)))
will get a method error, because of the way Javis.Rotation is defined
using a parametric type. Same deal for
    anim_rotate[_around](o1, <integer>, o2)
The small change in the definition of Rotation in this commit allows
the use of Irrational and Integer angles in anim_rotate_around()
and anim_rotate().